feat: aded findByCategory into ervice and Repository fix: make id Long into Movie

+++ src/main/java/com/pryect1/EscenaDelCrimen/service/MovieService.java
@@ -25,7 +25,7 @@ public class MovieService {
         return movieRepository.save(newMovie);
     }
 
-    public void deleteMovie(int id){
+    public void deleteMovie(Long id){
         movieRepository.deleteById(id);
     }
